pyproject.toml
[project]
name = "greenprompt"
version = "0.1.0"
description = ""
authors = [
    {name = "Uday",email = "udapra@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "openai (>=1.77.0,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "psutil (>=7.0.0,<8.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "py-cpuinfo (>=9.0.0,<10.0.0)",
    "flask-cors (>=6.0.0,<7.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
